+++ src/function/matrix/size.js
@@ -1,13 +1,14 @@
 import { arraySize } from '../../utils/array'
 import { factory } from '../../utils/factory'
-import { noMatrix } from '../../utils/noop'
 
 const name = 'size'
-const dependencies = ['typed', 'config', '?matrix']
+const dependencies = ['typed']
 
-export const createSize = /* #__PURE__ */ factory(name, dependencies, ({ typed, config, matrix }) => {
+export const createSize = /* #__PURE__ */ factory(name, dependencies, ({ typed }) => {
   /**
    * Calculate the size of a matrix or scalar.
+   *
+   * The function always returns an Array, also for Matrix input.
    *
    * Syntax:
    *
@@ -27,24 +28,22 @@ export const createSize = /* #__PURE__ */ factory(name, dependencies, ({ typed, 
    *     resize, squeeze, subset
    *
    * @param {boolean | number | Complex | Unit | string | Array | Matrix} x  A matrix
-   * @return {Array | Matrix} A vector with size of `x`.
+   * @return {Array} A vector with size of `x`.
    */
   return typed(name, {
     Matrix: function (x) {
-      return x.create(x.size())
+      return x.size()
     },
 
     Array: arraySize,
 
     string: function (x) {
-      return (config.matrix === 'Array') ? [x.length] : matrix([x.length])
+      return [x.length]
     },
 
     'number | Complex | BigNumber | Unit | boolean | null': function (x) {
       // scalar
-      return (config.matrix === 'Array')
-        ? []
-        : matrix ? matrix([]) : noMatrix()
+      return []
     }
   })
 })